+++ resources/assets/scripts/app.js
@@ -1,0 +1,9 @@
+/**
+ * External Dependencies
+ */
+import 'jquery';
+import 'bootstrap';
+
+$(document).ready(() => {
+  // console.log('Hello world');
+});