+++ resources/index.php
@@ -1,3 +1,0 @@
-<?php
-
-// this file is deliberately blank