+++ resources/views/components/alert.blade.php
@@ -1,0 +1,3 @@
+<div class="alert alert-{{ $type }}">
+  {!! $message ?? $slot !!}
+</div>