+++ resources/views/layouts/app.blade.php
@@ -1,23 +1,15 @@
-<!doctype html>
-<html {!! get_language_attributes() !!}>
-  @include('partials.head')
-  <body @php body_class() @endphp>
-    @php do_action('get_header') @endphp
-    @include('partials.header')
-    <div class="wrap container" role="document">
-      <div class="content">
-        <main class="main">
-          @yield('content')
-        </main>
-        @if (App\display_sidebar())
-          <aside class="sidebar">
-            @include('partials.sidebar')
-          </aside>
-        @endif
-      </div>
-    </div>
-    @php do_action('get_footer') @endphp
-    @include('partials.footer')
-    @php wp_footer() @endphp
-  </body>
-</html>
+@include('partials.header')
+
+<div class="container">
+  <main class="main">
+    @yield('content')
+  </main>
+
+  @hasSection('sidebar')
+    <aside class="sidebar">
+      @yield('sidebar')
+    </aside>
+  @endif
+</div>
+
+@include('partials.footer')