+++ resources/views/partials/entry-meta.blade.php
@@ -1,6 +1,10 @@
-<time class="updated" datetime="{{ get_post_time('c', true) }}">{{ get_the_date() }}</time>
+<time class="updated" datetime="{{ get_post_time('c', true) }}">
+  {{ get_the_date() }}
+</time>
+
 <p class="byline author vcard">
-  {{ __('By', 'sage') }} <a href="{{ get_author_posts_url(get_the_author_meta('ID')) }}" rel="author" class="fn">
+  <span>{{ __('By', 'sage') }}</span>
+  <a href="{{ get_author_posts_url(get_the_author_meta('ID')) }}" rel="author" class="fn">
     {{ get_the_author() }}
   </a>
 </p>