+++ resources/views/partials/sidebar.blade.php
@@ -1,1 +1,1 @@
-@php dynamic_sidebar('sidebar-primary') @endphp
+@php(dynamic_sidebar('sidebar-primary'))