+++ resources/views/search.blade.php
@@ -3,14 +3,15 @@
 @section('content')
   @include('partials.page-header')
 
-  @if (!have_posts())
-    <div class="alert alert-warning">
-      {{ __('Sorry, no results were found.', 'sage') }}
-    </div>
+  @if (! have_posts())
+    <x-alert type="warning">
+      {!! __('Sorry, no results were found.', 'sage') !!}
+    </x-alert>
+
     {!! get_search_form(false) !!}
   @endif
 
-  @while(have_posts()) @php the_post() @endphp
+  @while(have_posts()) @php(the_post())
     @include('partials.content-search')
   @endwhile