+++ resources/views/single.blade.php
@@ -1,7 +1,7 @@
 @extends('layouts.app')
 
 @section('content')
-  @while(have_posts()) @php the_post() @endphp
-    @include('partials.content-single-'.get_post_type())
+  @while(have_posts()) @php(the_post())
+    @includeFirst(['partials.content-single-' . get_post_type(), 'partials.content-single'])
   @endwhile
 @endsection